Allow customers to add payment methods if the toggle is enabled
If the store doesn't accept automatic renewals but has enabled the toggle
assume that they have turned them off by default but want customers to be
able to turn on automatic renewals if they want.

To make this possible we have to allow customers to add a payment method
if the store doesn't accept automatic renewals but has enabled the toggle.