Fix uncancel conditions
Previously, any subscription with the pending-cancellation status
could be changed to active. This would lead to issues with payment
gateways that couldn't support that status transition.

We have no 'uncancel' supports flag to check against, but we can
determine with sufficient certainty whether a payment gateway will
be able to uncancel the subscription based on other supports flags.