Don't display resubscribe when activation possible
With the introduction of #2993, it became possible for a customer
to uncancel their subscription (via the reactivate button on their
My Account > View Subscription page).

However, the Resubscribe button was also being displayed. The
differences between the two actions are limited to subscriptions
with a length. In that case, one will continue the existing length
while the other will start a new subscription fresh. In other cases,
the ultimate benefit of the actions is identical, in which case,
it doesn't make sense to display both.

Example screenshots:
 * before this commit: http://pic.pros.pr/cf4e241cfff8
 * after this commit: http://pic.pros.pr/82f274002144