Calculate the last order time from last non-early renewal or parent
One of the scenarios this branch sets out to fix is miscalculations
when early renewals are involved. At its core, the problem is caused by
early renewals extending the current billing term which in effect
causes the number of days until the next payment to be skewed which
also throws out the old price per day calculation.

This branch has a number of fixes for this scenario, this commit is only
1 part. The other changes will come later.

In the end, the new approach is to calculate the total amount the customer
has paid in early renewal and switches since the last scheduled renewal
or parent and then divide that across the entire billing term.