Refactor the calculation of the number of days in the old cycle
Previously the number of days in the old cycle was always calculated from
the hardcoded number of days in the subscriptions billing cycle
(1 month = 30.4375) or the number of days between the next payment and the
last payment.

If the subscription is synced and still in its first period we would always
use the hardcode number of days as it can't calculate the correct number of days
with the dynamic (sort of) free trial period that synced products have.

The new approach, introduced with this commit, acts in the same way but
extends the logic to also use the hard-coded number of days if the subscription
is still on trial and if the customer hasn't paid anything since signup.

This may still be error prone and may need to be revisited as using the hardcoded
number of days leads to weird results and should be avoided where possible.