Introduce a new function to find the total paid since the last order
Note: the last order, in this case, is the last non-early renewal or
parent order.

This function is helpful to find how much a customer has paid for a given
item in early renewals and switches since the last scheduled renewal or
parent order.

This function also maps products back through multiple switches and so if a
product has switched from product A to product B, this function will find the
amount paid for product B up until the switch and then start looking for
product A in orders before the switch, for example.