Replace old recurring total with total paid for the current period
The $old_recurring_total variable was used to find the old price per day
and use to be the subscription's recurring amount or the amount paid in the
last related order - the amount the customer paid for the product in that order.

This approach fixed a situation where the subscription recurring amount was discounted
by an admin user recently and so the accurate amount paid by the customer was the last
order amount, not the subscription amount.

This method of finding the old price per day didn't account for early renewals, or
multiple prorated switch amounts which contribute to the total paid for the
current billing period.