Introduce a function to calculate the old price per day
The old price per day is the total paid / the days in the old cycle.
As has been mentioned in previous commits, with this branch this is:

The total paid since the last non-early renewal or parent order divided by
the total number of days since the last non-early renewal or parent order.

This fixes a couple of issues to do with early renewals. Because early
renewals extend the current term, you end up with a number of days in the old
cycle effectively double the billing period, depending on when you renew early.
Using this extended billing period with the last order amount made for some
incorrect switch calculations.

For example, if you have a $10/month product and you renew early on day 1,
the of number days in the current cycle is 2 months, and the last amount you
paid was $10. This means the old price per day would have been $10/~60. Not $10/30.

The new approach will find the total paid since the last parent or
non-early renewal ($20) and divide it by the number days between the last non-early
renewal or parent and the next payment date (~60). The result of that is $20/60 == $10/30.

This approach also works no matter when the customer renews early. If for example
you renewed 3 days before your next payment date, the old approach would have found the
old price per day to be $10/33. Now it will still be $20/60.

Essentially this new approach means the old price per day is more closely aligned
with what the actual product's price per day is.