Introduce new function to calculate the days in new cycle
Previously the days in the new cycle was always the hardcoded number
of days returned by wcs_get_days_in_cycle(). For monthly subscriptions,
this would be 30.4375.

Compared to the days in the old cycle which would be calculated by using
the last and next payment dates, this could lead to a mismatch. This
most evident in February where the number of days between the last and
next payment dates will be 28.

Because the days in the new cycle would be rounded up, this meant there
was a 3 day difference. This discrepancy would then fall through to the price
per day calculations which would then lead to the customer getting
more or fewer days then they should if the switched and then switched back.

Or could lead to incorrect upgrade costs.

To resolve this mismatch in the days in each cycle (new and old), this
commit introduces a new way to find the number of days in the new cycle.

It will simulate if the customer had purchased the new product on the last
payment date, what would be the next payment and then how many days between
those two dates.

This puts the two calculations in line with each other.

1 example:

If you purchase a monthly subscription on February 15th, your next
payment will be March 15th. In 2019, that's 28 days.

If you switch to another monthly product, the days in the new cycle
would have been 30.4375. But with this new approach, it will be 28 too
because it will find the new product's next payment after February 15
which would be March 15.

The same will also be true in other months that have 30 or 31 days.

Previously the number there was a condition that would check if the
billing cycles matched, however this was no longer necessary.