Refactor the code which finds the new price per day
Prior to this commit, the new price per day was always the
new product's recurring price / the days in the new cycle.

This largely remains the way it's calculated in this commit, however
the one exception is when you're switching during a free trial where,
the trial periods match.

In this case we want the switch to continue the existing free trial and
so the new product price is $0.