Refactor the logic which finds if the prepaid term should be reduced
Previously switching would reduce the prepaid term if there were more days in the
old cycle as there are in the new cycle.

This logic still remains, however, this commit also adds an extra condition.

If the customer is switching during a free trial, and the trial periods don't match
we want to terminate the trial now and charge the full recurring price.

Because in this case, the customer wouldn't have paid anything, the code which
eventually reduces the prepaid term will find that the customer isn't entitled
to any days at the new price per day and so will start the subscription from today.