Introduce function to find the number of days to extend the term by
This part of the calculate_prorated_totals function changes a little from
the previous code. In the past, if the switch was a downgrade it would first
check if we needed to reduce the prepaid term and if it didn't, it would set
the $days_to_add variable to 0.

Further down this variable would then be multiplied by the number of seconds
in a day and so that too would also be 0.

This 0 value was then added to the $next_payment_timestamp and set to the switch
cart item's ['first_payment_timestamp']. At the very beginning of the switching
calculations, ['first_payment_timestamp'] is already set to the
$next_payment_timestamp and so adding 0 to that wouldn't have changed the result.

This wasn't necessary and so has been removed and simplified.