Refactor the code which calculates the upgrade cost
This logic largely remains unchanged other than to add a new condition to
ignore any length calculations while switching during a trial.

If a customer switches during a free trial where the new product has an end date,
they haven't paid anything and so we don't want for them to be charged the new price
per day for the number of days until the trial end date as this can lead to an exorbitant
upgrade cost. For an example see #2826. Further details about this change will be
included in the PR.