Apply deprecated filter when params match old function
WC_Subscription::get_completed_payment_count() use to apply the
woocommerce_subscription_payment_completed_count filter. That function
use to only find the completed count and it use to be a combined parent
and renewal count. Therefore when the new function (get_payment_count()),
is called and is finding the completed count for renewal and parent orders,
apply the filter.

Any other combination of params is new functionality and so doesn't need to
worry about the deprecated behaviour.