Add the item tax when manually renewing and the prices include tax
When the store's prices are inclusive of tax, we need to tell the cart
what the product's price is. In the past we've done this by using the
line item's line_subtotal (price not including tax) and then add
the current tax rates that apply.

Under normal circumstances this is fine since the line_subtotal + the current
tax amount will result in the same amount the customer initially paid.

If the tax rate change, however, the customer will pay more or less. This isn't
right.