Fix compatibility issue with WC Services
WC Services calculate the cart object totals twice. Our limited payment
coupon code was assuming that the coupon discount totals on a recurring $cart
object  were still the amounts cloned from the initial/global WC cart.

This led to some incompatibility issues.

To fix that, this commit ensures we use the WC()->cart when
we want to check the initial cart discount totals.