Round the calculated days between the last and next payment
Previously we would floor the result which led to a situation where the number of days might be 183.99998842593
which is less than a second of being 184 but it is rounded down to 183. When you then compare that
to a days in old cycle which more correctly finds it to be 184, you can get weird upgrade costs.

I have a feeling this will then lead to other problems howwever. If it does, we'll need to find
a smarter way of comparing the number of days in the new and old cycle.

Issue for reference: https://github.com/Prospress/woocommerce-subscriptions/issues/3390#issuecomment-514901415