Always load the wcs_get_last_non_early_renewal_order function
The function was located in the early-renewal/wcs-early-renewal-functions.php
file which is only loaded when the early renewal feature is enabled.

An alternative fix would be to always load the early renewal functions
file, however, that isn't possible because it protected behind a check
to see if the store is running the mini-extension version of the feature.