Only update the report cache once and prevent WC overriding our cache
WC 3.7 now only updates the report cache on shutdown. The way it has been
implemented means they only update it with data generated by get_order_report_data().

Because we only generate a couple of figures using that method, WC 3.7 was overriding the
cache we were updating, with only a subset of the query results.