Only filter the product class for products which still exist
WC 3.6 added a caching layer to the function which gets the product type
from the post type. WC_Product_Data_Store_CPT::get_product_type().

This means that even if a product is deleted/trashed, WC may return a
type from the cache rather than what would have been false pre WC 3.6.

This maybe a bug or a breaking change in WC core. In any case, the fix on our end
is to check the post exists before trying to use it.