Honour the default value of the defer emails hook
When this hook was introduced in WC core initially, the default was true,
they changed it later to be false by default.

This was casing 2 emails to be sent. I don't know if this was a real
possibility on live sites since this function is only used in the retry system,
and it would have reattached the emails on the deferred emailer.

So only emails sent after the retry system had handled blocking the
admin order emails would emails have been affected. By default only the
failed order email is sent in those cases.