Introduce function to allow filtering an array using a WP filter
To do this we had to deprecate the $carts_with_multiple_payments variable because it's no longer
reliable. It use to be used to set the various sections's row span, however we can now use the number of
recurring carts displaying a total instead.

I also discovered a $shipping_methods variable that's no longer used too.