Add a check for on-hold status when canceling sub
When a user manually cancels a subscription, we perform some checks
before assigning a relevant status.

The first is to check if the subscription is currently active and
whether there is still prepaid time. In this case, we set the
subscription to pending-cancel. This check does not account for
subscriptions that are placed on-hold as a result of early renewals.

Adding a on-hold to this first conditional catches these cases.