Add cancelled status to needs_payment check
This adds a check for whether a subscriptions parent order is cancelled
when deciding if the subscription needs payment. This is to account for
cases when the user cancels the parent order for an on-hold sub.

Previously, the subscription would be placed pending-cancel if needs_payment
returned false, but adding this check ensures needs_payment returns true
for the above case.