Return false when signup date is today
The `is_payment_upfront()` method is used by
`WC_Subscription_Product::get_price_string()` to decide whether to
prepend the price string with '$X now`.

When signup date is today, this leads to some confusion as the price
string also includes the subscription length. So it appears as if the
subscription requires one additional payment. In cases where signup is
today, we instead want to return false as technically there is no
upfront (prorated) payment. We are actually just starting the
subscription.

This commit does that, returning false when signup date is today.