Remove redundant subscription status update
This removes what appears to be a redundant call to update_status() on
subscription when processing IPN requests.

A preceding call to $order->process_complete() eventually triggers a
different call to update_status(). When both of these are called, the
subscription in question ends up with two status changed notes.

Removing this bit of code avoids this duplication.