Handle update order meta from blocks
This is because the WooCommerce Blocks extension does actions differently to core, therefore we need to handle these actions in the same way we would handle core actions. We have updated the signatures for the callback methods to make $posted_data optional, as it is not used, nor is it passed by the WooCommerce Blocks __experimental_woocommerce_blocks_checkout_update_order_meta action.