Deprecate WC_Subscriptions::attach_dependant_hooks()
The last remaining functionality of attach_dependant_hooks was to
tell the autoloader whether to load the legacy versions of the API
because we dropped support for WC pre 3.0 that was always going to
be false. That also happens to be the default value of the autoloaders
legacy_api property. The setting of that property is no longer necessary.

See: https://github.com/woocommerce/woocommerce-subscriptions/blob/introduce/base-plugin-class/includes/class-wcs-autoloader.php#L285
See: https://github.com/woocommerce/woocommerce-subscriptions/blob/introduce/base-plugin-class/includes/class-wcs-autoloader.php#L27