Allow Subscriptions Core to be activated as a stand-alone plugin
Having subscriptions-core activated as a plugin will take priority over subscriptions-core inside subscriptions and woocommerce payments.
This is mostly to make developing and testing the package much easier.

Activating and deactiving this on its own, doesn't do anything. This is intended because
WC Subscriptions and WC Payments will be checking if the plugin is activated to determine where to load subscriptions-core from