Move checkout blocks loading to init rather than 'plugins_loaded'
Subscriptions previously use to have the blocks integration code loaded
at the top of the plugin file. When we split the main plugin file into
the two plugin classes it incorrectly got moved into the function that
runs on plugin loaded, which is too late for the blocks hook as it has
alread ran.

This commit fixes that issue by making sure we attach the block hooks
on init.