Update the subscription's payment method when changing the default one for all
This commit updates some extra metadata when updating a subscription token. We're now adding the newly selected token to the subscription's meta that keeps track of the token list, and we're updating the payment method for the subscription in case the newly added token belongs to a different payment gateway than the previous one.