Set the interval value to 1 and disable its select for yearly periods
Stripe doesn't allow having greater intervals for yearly subscriptions. We used to allow creating products with greater intervals making the checkout process fail without a clear message why it fails.

This commit disables the interval select and sets it to '1' when the product period is set to 'year', in the client side.