Deprecate the downgrade notice function
We added this downgrade admin notice to warn store managers that were downgrading to really old
versions of subscriptions.

Originally I was thinking we'd need to change this logic to only show this notice if the store downgraded to 1.5, but that doesn't make sense.
if the store downgraded to 1.5 of WC Subscriptions, they wouldn't have these changes...

The solution here is to just deprecate this function as it's of no use.