Use WC_Subscriptions_Product::is_subscription to determine if product is purchased on subscription
To convert any product to a subscription (regardless of type), APFS sets the current "subscription plan" on product instances directly, and filters the output of WC_Subscriptions_Product::is_subscription to help WCS recognize any product as a subscription.

WCS is coded in a type-agnostic manner to be compatible with APFS. In contexts where a product object is available, it should always be used instead of its id or type to detect whether it's a "subscription product" or not.