Update wcs_get_objects_property() to always use object->get_meta()
This change is to avoid ever using get_post_meta on an order
The reason for this is because with orders moving to custom tables, we never want
to fetch data from the postmeta table.

Removing the `get_post_meta()` calls also potentially  breaks this function being used for fetching post data,
hence why it's being kept.