Adds new COT compatibility function to fix fetching subscriptions
When custom order tables is enabled on a store, calling wc_get_orders() with a subscription ID
no longer works to get a subscription object. This is because Subscriptions are staying
in the WP post tables and the new COT data store is trying to get the shop subscription order type which doesn't exist.

This new function hooks onto the order classes filter and uses the subscriptions data store to check if the ID
belongs to a subscription, if so, return the WC Subscriptions class name