Deprecate the maybe_preserve_order_status() function
The maybe_preserve_order_status() function was wreaking havoc on the
default order status when the cart contained a renewal order with
a failed status.

This function was originally introduced to make sure the renewal order
kept its failed status as it was processed through the checkout. That
is no longer necessary. Order's do maintain their existing status.
This can be confirmed by observing the old order status in
WC_Subscriptions_Renewal_Order::maybe_record_subscription_payment().

The original commit this function was introduced in can be found here:
2cc025d887093720f4134d7cc411ddf093d3120b