Added break statements in switch (#533)
Missing break statements inside the switch ... case block in
WC_Subscription::validate_date_updates() are causing an unexpected
behavior when validating dates while changing subscription status.

According to PHP documentation, when a case statement is found whose
expression (...) matches the value of the switch expression (...) PHP
continues to execute the statements until the end of the switch block.

So for example, when the case matches 'end' then all the following
lines inside the block will be executed as if the expression would
match 'cancelled', 'next_payment' and 'trial_end'.